shim: Add an alternate way to communicate the workload exit status
While it's totally natural to take the workload exit status from reaping
the process, with a runtime using VMs to launch "containers" we end up
reaping the exit status of the hypervisor (eg. qemu).

We could have yet another process around the runtime just to take care
of the communication between the hypervisor and containerd-shim: that
other process would get the workload exit status and return it with
exit() and we'd get containerd-shim to reap that process instead of the
hypervisor.

However, this introduce a per-container cost, which isn't great.

We can just have containerd-shim offering a pipe for runtimes to write
the exit status into. If the runtime so chooses, it can use that pipe
and container-shim will use the value read from it instead of the reaped
exit status. If nothing is written into that pipe, the usual exit status
prevails.

Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>